backend/requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0          # Azure App Service brug

# ── Database ───────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.10
sqlalchemy==2.0.30
alembic==1.13.1           # Databasemigrationer

# ── Validation ─────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Auth (Azure Entra ID JWT-validering) ──────────────────────────────────────
python-jose[cryptography]==3.3.0
httpx==0.27.0             # til JWKS-hentning fra Microsoft

# ── Magic link (brobygger-login) ─────────────────────────────────────────────
itsdangerous==2.2.0       # signerede tokens til magic links

# ── Email (magic links + notifikationer via Exchange Online) ─────────────────
aiosmtplib==3.0.1

# ── Real-time (SignalR) ────────────────────────────────────────────────────────
azure-signalr==1.1.0

# ── Web Push / VAPID ──────────────────────────────────────────────────────────
pywebpush==2.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-multipart==0.0.9   # form data
python-dotenv==1.0.1

# ── Test ──────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0             # TestClient til FastAPI
